pass test codeLab3

+++ src/main/java/org/adaschool/api/controller/product/ProductsController.java
@@ -1,6 +1,8 @@
 package org.adaschool.api.controller.product;
 
+import org.adaschool.api.exception.ProductNotFoundException;
 import org.adaschool.api.repository.product.Product;
+import org.adaschool.api.repository.product.ProductDto;
 import org.adaschool.api.service.product.ProductsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -23,35 +25,58 @@ public class ProductsController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Product> findById(@PathVariable String id){
-        try{
-            return new ResponseEntity(productsService.findById(id), HttpStatus.OK);
-        }catch (NoSuchElementException e){
-            return new ResponseEntity("the producto " + id + "does't in the data base." , HttpStatus.NOT_FOUND);
+    public ResponseEntity<Product> findById(@PathVariable String id) {
+        try {
+            Optional<Product> product = productsService.findById(id);
+
+            if (product.isPresent()) {
+                return ResponseEntity.ok(product.get());
+            } else {
+                throw new ProductNotFoundException(id);
+            }
+        } catch (NoSuchElementException e) {
+            throw new ProductNotFoundException(id);
         }
     }
+
 
     @PostMapping
     public ResponseEntity<Product> save(@RequestBody Product product){
         return new ResponseEntity<>(productsService.save(product),HttpStatus.CREATED);
     }
 
-    @PutMapping("/{idProduct}")
-    public ResponseEntity<Product> updateProduct(@PathVariable String id, @RequestBody Product product){
-        try{
-            Optional<Product> isUpdateProduct = Optional.ofNullable(productsService.update(product, id));
-            if(isUpdateProduct.isPresent()){
-                return new ResponseEntity("the product update is ok.", HttpStatus.OK);
-            }else {
-                return new ResponseEntity(HttpStatus.NOT_FOUND);
-            }
-        }catch (NoSuchElementException e){
-            return new ResponseEntity("the product " + id + " does't in the data base.", HttpStatus.NOT_FOUND);
+    @PutMapping("{id}")
+    public ResponseEntity<Product> update(@PathVariable("id") String id, @RequestBody ProductDto productDto) {
+        Optional<Product> optionalUser = productsService.findById(id);
+
+        if (optionalUser.isPresent()) {
+            Product existinProduct = optionalUser.get();
+            existinProduct.setName(productDto.getName());
+            existinProduct.setDescription(productDto.getDescription());
+            existinProduct.setCategory(productDto.getCategory());
+            existinProduct.setPrice(productDto.getPrice());
+
+            productsService.save(existinProduct);
+
+            return ResponseEntity.ok().build();
+        }
+        else {
+            throw  new ProductNotFoundException(id);
         }
     }
 
-    @DeleteMapping("/{id}")
-    public void delete(@PathVariable String id){
+
+    @DeleteMapping("{id}")
+    public ResponseEntity<Void> deleteProduct(@PathVariable("id") String id) {
+        Optional<Product> existingProduct = productsService.findById(id);
+
+        if (existingProduct.isEmpty()) {
+            throw new ProductNotFoundException(id);
+        }
+
         productsService.deleteById(id);
+        return ResponseEntity.ok().build(); // Cambiado de noContent() a ok()
     }
+
+
 }
